Feat: Proteção JWT, Admin e respostas na API

+++ server/src/college/college.service.ts
@@ -1,4 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { College } from '../entities/college.entity';
@@ -6,6 +10,7 @@ import { ValidEmail } from '../entities/valid-email.entity';
 import { CreateCollegeDto } from './dto/create-college-dto';
 import { CreateValidEmailDto } from './dto/create-valid-email-dto';
 import { Company } from '../entities/company.entity';
+import { RequestWithUser } from '../auth/request-with-user.interface';
 
 @Injectable()
 export class CollegeService {
@@ -18,8 +23,12 @@ export class CollegeService {
     private readonly companyRepository: Repository<Company>,
   ) {}
 
-  async createCollege(createCollegeDto: CreateCollegeDto) {
+  async createCollege(
+    createCollegeDto: CreateCollegeDto,
+    req: RequestWithUser,
+  ) {
     const { company_id } = createCollegeDto;
+    const userId = req.user.userId;
 
     const company = await this.companyRepository.findOne({
       where: { id: company_id },
@@ -27,12 +36,18 @@ export class CollegeService {
     });
 
     if (!company) {
-      throw new Error('Company not found');
+      throw new NotFoundException('Empresa não encontrada.');
     }
 
     if (company.user.type !== 'college') {
-      throw new Error(
-        'The user associated with this company must be of type "college".',
+      throw new UnauthorizedException(
+        'O usuário associado a esta empresa deve ser do tipo "faculdade".',
+      );
+    }
+
+    if (company.user.id !== userId) {
+      throw new UnauthorizedException(
+        'Usuário não autorizado a realizar esta ação.',
       );
     }
 
@@ -40,31 +55,104 @@ export class CollegeService {
     return this.collegeRepository.save(college);
   }
 
-  async createValidEmail(createValidEmailDto: CreateValidEmailDto) {
+  async createValidEmail(
+    createValidEmailDto: CreateValidEmailDto,
+    req: RequestWithUser,
+  ) {
     const { domain, college_id } = createValidEmailDto;
+    const userId = req.user.userId;
 
     const college = await this.collegeRepository.findOne({
       where: { id: college_id },
+      relations: ['company'],
     });
+
     if (!college) {
-      throw new Error('College not found');
+      throw new NotFoundException('Faculdade não encontrada.');
+    }
+
+    const company = await this.companyRepository.findOne({
+      where: { id: college.company.id },
+      relations: ['user'],
+    });
+
+    if (company.user.id !== userId) {
+      throw new UnauthorizedException(
+        'Usuário não autorizado a realizar esta ação.',
+      );
     }
 
     const validEmail = this.validEmailRepository.create({ domain, college });
     return this.validEmailRepository.save(validEmail);
   }
 
-  async deleteValidEmail(id: number) {
+  async deleteValidEmail(id: number, req: RequestWithUser) {
+    const userId = req.user.userId;
+
+    const validEmail = await this.validEmailRepository.findOne({
+      where: { id },
+      relations: ['college'],
+    });
+
+    if (!validEmail) {
+      throw new NotFoundException('E-mail válido não encontrado.');
+    }
+
+    const company = await this.companyRepository.findOne({
+      where: { id: validEmail.college.company.id },
+      relations: ['user'],
+    });
+
+    if (company.user.id !== userId) {
+      throw new UnauthorizedException(
+        'Usuário não autorizado a realizar esta ação.',
+      );
+    }
+
     return this.validEmailRepository.delete(id);
   }
 
-  async listValidEmails(college_id: number) {
+  async listValidEmails(college_id: number, req: RequestWithUser) {
+    const userId = req.user.userId;
+
+    const college = await this.collegeRepository.findOne({
+      where: { id: college_id },
+      relations: ['company'],
+    });
+
+    if (!college) {
+      throw new NotFoundException('Faculdade não encontrada.');
+    }
+
+    const company = await this.companyRepository.findOne({
+      where: { id: college.company.id },
+      relations: ['user'],
+    });
+
+    if (company.user.id !== userId) {
+      throw new UnauthorizedException(
+        'Usuário não autorizado a realizar esta ação.',
+      );
+    }
+
     return this.validEmailRepository.find({
       where: { college: { id: college_id } },
     });
   }
 
-  async listAllValidEmails() {
+  async listAllValidEmails(req: RequestWithUser) {
+    const userId = req.user.userId;
+
+    const userColleges = await this.collegeRepository.find({
+      where: { company: { user: { id: userId } } },
+    });
+
+    if (userColleges.length === 0) {
+      throw new UnauthorizedException(
+        'Usuário não autorizado a visualizar os e-mails.',
+      );
+    }
+
     return this.validEmailRepository.find();
   }
 }
